Add changes required for packer; amend kops role to properly build repo
The subnet id and vpc is required for the packer AMI to build in the correct VPC and in the private subnet. It is not built with a public IP so requires a private subnet with NAT gateway access.

The kops repository was not getting built correctly and required a rebuild of the yum cache to work correctly.